LoginPage now gets login info from context
