+++ frontend/src/shared/urls.js
@@ -1,0 +1,3 @@
+export const BACKEND_HOST = "http://localhost:3000"
+export const SIGN_IN_URL = `${BACKEND_HOST}/users/sign_in`
+export const CHECK_LOGGED_IN_URL = `${BACKEND_HOST}/test_login/is_logged_in`
